requirements.txt
torch
sklearn
pandas
matplotlib
pytorch-lightning==1.6.5
torch-optimizer
timm
wandb
hydra-core
hydra-colorlog
scikit-image
scikit-learn
python-dotenv
rich
einops
paramiko
mat73
pytorch-lightning-bolts
torchmetrics==0.9.2
albumentations
scp 
transformers